requirements.txt
streamlit
pandas
gspread
google-auth
requests
python-dotenv
openai
groq